feat(motionplayer): M2 — EmoteEngine transition/deque#8 vertical (+ EmoteVarController float-bits fix)
Fifth M2 stepped-deque vertical (5/6 stepped deques now live). Faithful port per
fresh decompile of EmoteVarController_step sub_666BF8, buildTransitionControl
@0x66D4C4, EmoteVarController_ctor@0x667030, and Animator_setKeyframes@0x667300.

- buildTransitionControl: enabled gate, new(0x80) EmoteVarController(count=1)+ctor,
  push 24B {ctl@+0, label@+8, flag@+16=1} into deque@+576 (_auxVarDeque8), HM@+1384
  {type=7, index}. progress() deque#8 STUB_WARN replaced with the step + HM7[label]
  upsert (24B stride), ordered after selector; dtor destroys it.
- EmoteVarController_step: 2-state ramp (state0 latch target/start from keyframe +
  invDuration=1/elem[+12]; state1 phase+=invDur*dt, current=target+powf(phase,pow)*
  (start-target), phase>=1 -> snap + state0). powCount = elem[+16] raw DWORD copy.

float-bits fix (the known bug flagged by the selector slice): EmoteVarController
read the 20B keyframe powCount as static_cast<float>(int) (uint32 field), but
sub_666BF8 loads it via LDR S1 @0x666df4 (no SCVTF) = raw float bits. Changed
EmoteVarKeyValue20B::powCount and EmoteVarController::powCount uint32/int32 ->
float, store via memcpy, use directly in std::pow. Also fixed
EmoteSelectorController.cpp:67 memcpy size + stale "uint32" comment.

Dispatch order: PlayerCore applyMetadata now builds transitionControl BEFORE
selectorControl (matching 0x67D4D0), so selector options resolve their refCtl
cross-references into the transition deque (selector refCtl scan is now live).

Build verified green independently (web/debug exit 0 grep error: = 0; wasmtime
guest exit 0, no undefined/unsupported import; no residual EmoteAuxVar24B_Deque8).
Self-audited line-by-line vs disasm (float-bits write/copy/use all raw-bit,
no SCVTF).

Open in M2: deque#10 inline curve (progress @0x67d2a0), bust/hair/parts physics,
sub_661F7C resolver, setVariable cases 4-8 reader dispatch, clamp/mirror/loop/
instantVariable builders, and EmoteAngleController.cpp:85/105 (same-class
powCount int-cast bug, next).

Not runtime-verified: no oracle/fixture exercises transitionControl/powCount; per
CLAUDE.md no fixture fabricated; verification = build + disasm-level alignment.

+++ cpp/plugins/motionplayer/EmoteSelectorController.cpp
@@ -59,12 +59,14 @@ namespace motion {
                 kf.channel[i] = values[i];
             }
             kf.duration = dur;
-            // pow is a float in the binary keyframe element's powCount slot. The
-            //   element field is uint32 powCount but it is read back with LDR S
-            //   (raw float bits) in EmoteVarController_step, so store the float
-            //   bits directly (consistent with the mouth/eye/eyebrow powField
-            //   raw-bits rule — Animator_setKeyframes receives a float `pow`).
-            std::memcpy(&kf.powCount, &pow, sizeof(uint32_t));
+            // pow is a float in the binary keyframe element's powCount slot
+            //   (+16). The whole pipeline treats +16 as raw float bits: it is
+            //   written by Animator_setKeyframes (0x667300) via a DWORD copy of a
+            //   float arg, and read back with `LDR S1` (no SCVTF) in
+            //   EmoteVarController_step (0x666df4). EmoteVarKeyValue20B::powCount
+            //   is now a `float`, so this memcpy is a plain raw-bit store of the
+            //   float `pow` (consistent with the mouth/eye/eyebrow powField rule).
+            std::memcpy(&kf.powCount, &pow, sizeof(float));
             ctl->queue.push_back(kf);
         }
